refactor(ui): useSocial hooks refactor
useSOcial hooks refactor
