+++ packages/ui/src/containers/SocialSignIn/PrimarySocialSignIn/index.tsx
@@ -1,5 +1,7 @@
 import React from 'react';
 
+import useSocialSignInListener from '@/hooks/use-social-signin-listener';
+
 import SocialSignInList from '../SocialSignInList';
 
 export const defaultSize = 3;
@@ -9,6 +11,8 @@ type Props = {
 };
 
 const PrimarySocialSignIn = ({ className }: Props) => {
+  useSocialSignInListener();
+
   return <SocialSignInList className={className} />;
 };
